third_party/blink/web_tests/external/wpt/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-namespace-expected.txt

This is a testharness.js-based test.
FAIL getElementsByName and foreign namespaces assert_equals: expected 1 but got 2
Harness: the test ran to completion.
